# modules/apps/headless/headless-search/headless-search-impl/rest-openapi.yaml
components:
    schemas:
        Creator:
            description: https://www.schema.org/Creator
            properties:
                additionalName:
                    type: string
                familyName:
                    type: string
                givenName:
                    type: string
                id:
                    format: int64
                    type: integer
                image:
                    format: uri
                    type: string
                name:
                    type: string
                profileURL:
                    format: uri
                    type: string
            type: object
        Document:
            description: https://www.schema.org/Document
            properties:
                resourceType:
                    type: string
                author:
                    type: string
                clicks:
                    type: string
                date:
                    type: string
                description:
                    type: string
                hidden:
                    type: string
                id:
                    type: string
                pinned:
                    type: boolean
                title:
                    type: string
                type:
                    type: string
            type: object
        SearchResult:
            description: https://www.schema.org/SearchResult
            properties:
                resourceType:
                    type: string
                documents:
                    items:
                        $ref: "#/components/schemas/Document"
                    type: array
                total:
                    format: int64
                    type: integer
            type: object
info:
    description: ""
    title: "Headless Search"
    version: v1.0
openapi: 3.0.1
paths:
    "/document/{index}/{uid}":
        get:
            operationId: getDocument
            parameters:
                - in: path
                  name: index
                  required: true
                  schema:
                      type: string
                - in: path
                  name: uid
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Document"
                    description: ""
            tags: ["Document"]
    "/search/hidden/{companyId}/{index}/{keywords}/{from}/{size}":
        get:
            operationId: getSearchHiddenCompanyIndexKeywordsFromSize
            parameters:
                - in: path
                  name: companyId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: index
                  required: true
                  schema:
                      type: string
                - in: path
                  name: keywords
                  required: true
                  schema:
                      type: string
                - in: path
                  name: from
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: size
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/SearchResult"
                    description: ""
            tags: ["SearchResult"]
    "/search/{companyId}/{keywords}/{from}/{size}":
        get:
            operationId: getSearchCompanyKeywordsFromSize
            parameters:
                - in: path
                  name: companyId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: keywords
                  required: true
                  schema:
                      type: string
                - in: path
                  name: from
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: size
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/SearchResult"
                    description: ""
            tags: ["SearchResult"]